feat(#91): refactor: crawl4ai.ts — audit and modularize
